+++ client/core/src/com/cows/game/controllers/TileController.kt
@@ -1,6 +1,6 @@
 package com.cows.game.controllers
 
-import com.cows.game.Coordinate
+import com.cows.game.map.Coordinate
 import com.cows.game.enums.TileType
 import com.cows.game.models.TileModel
 import com.cows.game.views.TileView
