+++ client/core/src/com/cows/game/controllers/Updatable.kt
@@ -1,7 +1,6 @@
 package com.cows.game.controllers
 
-import com.cows.game.Updater
-import com.cows.game.views.Renderable
+import com.cows.game.managers.Updater
 
 abstract class Updatable {
     abstract fun update(deltaTime:Float)
