+++ client/core/src/com/cows/game/models/TileModel.kt
@@ -1,6 +1,6 @@
 package com.cows.game.models
 
-import com.cows.game.Coordinate
+import com.cows.game.map.Coordinate
 import com.cows.game.enums.TileType
 
 data class TileModel(
