+++ client/core/src/com/cows/game/models/TowerModel.kt
@@ -1,6 +1,6 @@
 package com.cows.game.models
 
-import com.cows.game.Coordinate
+import com.cows.game.map.Coordinate
 import com.cows.game.enums.TowerType
 
 data class TowerModel (
