+++ client/core/src/com/cows/game/views/Renderable.kt
@@ -1,7 +1,7 @@
 package com.cows.game.views
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
-import com.cows.game.Renderer
+import com.cows.game.managers.Renderer
 
 abstract class Renderable {
     abstract fun render(batch: SpriteBatch, deltaTime: Float)
