+++ client/core/src/com/cows/game/views/UnitView.kt
@@ -3,10 +3,8 @@ package com.cows.game.views
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.graphics.g2d.Sprite
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
-import com.cows.game.Coordinate
-import com.cows.game.enums.TileType
+import com.cows.game.map.Coordinate
 import com.cows.game.enums.UnitType
-import com.cows.game.models.TileModel
 import com.cows.game.models.UnitModel
 
 class UnitView (val model: UnitModel): Renderable() {
